requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-decouple==3.8
celery==5.3.4
redis==5.0.1
drf-yasg==1.21.7
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
freezegun==1.2.2
django-cors-headers==4.3.1
gunicorn==21.2.0
